app.js setup
